+++ History.md
@@ -1,7 +1,14 @@
+0.1.0 / 2017-02-06
+==================
+
+  * `.xml()` assertion to test res Content-Type to be application/xml
+  * `.contentType(type, charset)` now used for every content type assertions
+  * formatter for http.IncomingMessage
+
 0.0.3 / 2015-05-21
 ==================
-  
-  * .status will show only body and statusCode
+
+  * `.status()` will show only body and statusCode
 
 0.0.2 / 2014-05-29
 ==================